delete FileWidgetUtils and MediaUtil
