+++ collect_app/src/main/java/org/odk/collect/android/widgets/AudioWidget.java
@@ -30,12 +30,12 @@ import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.audio.AudioControllerView;
 import org.odk.collect.android.databinding.AudioWidgetAnswerBinding;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
+import org.odk.collect.android.utilities.ContentUriProvider;
 import org.odk.collect.android.utilities.QuestionMediaManager;
 import org.odk.collect.android.utilities.WidgetAppearanceUtils;
 import org.odk.collect.android.widgets.interfaces.WidgetDataReceiver;
 import org.odk.collect.android.widgets.utilities.AudioFileRequester;
 import org.odk.collect.android.widgets.utilities.AudioPlayer;
-import org.odk.collect.android.widgets.utilities.FileWidgetUtils;
 import org.odk.collect.android.widgets.utilities.RecordingRequester;
 import org.odk.collect.audioclips.Clip;
 
@@ -126,7 +126,7 @@ public class AudioWidget extends QuestionWidget implements WidgetDataReceiver {
                 questionMediaManager.replaceAnswerFile(getFormEntryPrompt().getIndex().toString(), newAudio.getAbsolutePath());
 
                 Uri audioURI = getContext().getContentResolver().insert(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI,
-                        FileWidgetUtils.getContentValues(newAudio, false));
+                        ContentUriProvider.getContentValues(newAudio, false));
 
                 if (audioURI != null) {
                     Timber.i("Inserting AUDIO returned uri = %s", audioURI.toString());
@@ -148,7 +148,6 @@ public class AudioWidget extends QuestionWidget implements WidgetDataReceiver {
             }
         } else {
             Timber.w("AudioWidget's setBinaryData must receive a File object.");
-            return;
         }
     }
